tomas_premoli/requirements.txt
django 
djangorestframework 
django-cleanup
pillow
requests
zappa
mysqlclient
boto3
django-storages